ENSO/app0-app5/requirements.txt
hopeit.engine[web,cli,fs-storage,redis-streams,redis-storage,apps-visualizer,config-manager,apps-client]>=0.19.0,<0.20.0
app0.admin==0.3.7
app0.platform.auth==0.3.7
motor[srv]>=2.5.1
bcrypt>=4.0.1
jinja2>=3.1.1
aioboto3>=9.6.0
paho-mqtt>=1.6.1
pendulum>=2.1.2
aiomqtt>=1.2.1
pandas>=2.0.3
ortools>=9.6.2534
aiocsv>=1.2.4